+++ accumulo/src/main/java/org/vertexium/accumulo/AccumuloVertex.java
@@ -354,6 +354,36 @@ public class AccumuloVertex extends AccumuloElement implements Vertex {
     }
 
     @Override
+    public EdgesSummary getEdgesSummary(Authorizations authorizations) {
+        Map<String, Integer> outEdgeCountsByLabels = new HashMap<>();
+        Map<String, Integer> inEdgeCountsByLabels = new HashMap<>();
+
+        if (inEdges instanceof EdgesWithCount) {
+            EdgesWithCount edgesWithCount = (EdgesWithCount) this.inEdges;
+            inEdgeCountsByLabels.putAll(edgesWithCount.getEdgeCountsByLabelName());
+        } else {
+            for (Map.Entry<Text, org.vertexium.accumulo.iterator.model.EdgeInfo> entry : getEdgeInfos(Direction.IN)) {
+                String label = entry.getValue().getLabel();
+                Integer c = inEdgeCountsByLabels.getOrDefault(label, 0);
+                inEdgeCountsByLabels.put(label, c + 1);
+            }
+        }
+
+        if (outEdges instanceof EdgesWithCount) {
+            EdgesWithCount edgesWithCount = (EdgesWithCount) this.outEdges;
+            outEdgeCountsByLabels.putAll(edgesWithCount.getEdgeCountsByLabelName());
+        } else {
+            for (Map.Entry<Text, org.vertexium.accumulo.iterator.model.EdgeInfo> entry : getEdgeInfos(Direction.OUT)) {
+                String label = entry.getValue().getLabel();
+                Integer c = outEdgeCountsByLabels.getOrDefault(label, 0);
+                outEdgeCountsByLabels.put(label, c + 1);
+            }
+        }
+
+        return new EdgesSummary(outEdgeCountsByLabels, inEdgeCountsByLabels);
+    }
+
+    @Override
     public Iterable<Vertex> getVertices(Direction direction, Authorizations authorizations) {
         return getVertices(direction, getGraph().getDefaultFetchHints(), authorizations);
     }