+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Changed: FetchHints to support more filtering
 * Changed: Accumulo: Store metadata in an indexed list to prevent duplication in memory and over the wire 
 * Fixed: Accumulo Iterator memory leak 
+* Added: Method to Vertex to get more detailed edge summary
 
 # v3.2.2
 * Fixed: Improved support for multithreaded clients with InMemoryGraph